Updated Inventory
WASD to move.
Once over the hydron press E to pick it up.
Press the red box in UI inventory to drop the item. Move around a bit so they don't stack.

To create the prefab, create a UI image and add the image of the item you want to be displayed into the inventory to this. Then make the actual item a prefab and place this in the game.

If you look at my scene and basiclly copy that is should work.