Add files via upload
If sizes of the transparent space behind the words need to be cropped let me know, also there are two quits cus I wanted to see how it would look if the sprite changes when clicked/hovered over. I also have some menu screens for the pause menu but I'll make a variety so we can pick and choose depending on the layout.  (Also i made the font so if you want anything specific made let me know too)